Refactor ExploreController
